Refactor to allow single JVM tests
