+++ common/src/main/java/org/cluster/membership/common/model/util/EnvUtils.java
@@ -1,0 +1,56 @@
+package org.cluster.membership.common.model.util;
+
+import java.io.File;
+import java.net.Socket;
+import java.net.SocketException;
+
+import org.cluster.membership.common.model.Node;
+import org.springframework.boot.system.ApplicationHome;
+
+public class EnvUtils {
+
+	public static boolean isAvailable(String host, int port) {
+		try {
+			(new Socket(host, port)).close();
+			return false;
+		}
+		catch(SocketException e) { return true; } 
+		catch(Exception e) { return false; }		  
+	}
+	
+	public static boolean isListening(String host, int port) {
+		try {
+			(new Socket(host, port)).close();
+			return true;
+		}
+		catch(SocketException e) { return false; } 
+		catch(Exception e) { return false; }		  
+	}
+
+	public static int getLocalAvailablePort(int port) {
+		for(int i = port; i < Short.MAX_VALUE - 1; i++) {
+			if(isAvailable("localhost", i)) return i;
+		}
+		return -1;
+	}
+	
+
+	public static String getHomePath(Class<?> clasz, String configFolderName) {
+		ApplicationHome home = new ApplicationHome(clasz); 
+		String path = home.getDir().getAbsolutePath();
+		String configFolderPath = path + File.separator + configFolderName;
+		File configFolder = new File(configFolderPath);
+		if(!configFolder.exists()) return path + File.separator + "target";
+		return path;
+	}
+
+	public static String generateNodeCommandLineArguments(Node node, int idx) {
+		return " --id." + idx + "=" + node.getId() +
+				" --address." + idx + "=" + node.getAddress() +
+				" --protocol.port." + idx + "=" + node.getProtocolPort() +
+				" --server.port." + idx + "=" + node.getServicePort() +
+				" --time.zone." + idx + "=" + node.getTimeZone().getID();
+	}
+
+	
+}
