+++ common/src/main/java/org/cluster/membership/common/model/util/MathOp.java
@@ -4,12 +4,9 @@ public class MathOp {
 
 	public static int log2n(int n) {
 		int base = 2;
-		int iterations = 0;
-		
-		while((base*=2) <= n*2) iterations++;
-		
-		return iterations + 1;
-		
+		int iterations = 0;		
+		while((base*=2) <= n*2) iterations++;		
+		return iterations + 1;		
 	}
 	
 }
