+++ protocol/src/test/java/org/cluster/membership/protocol/ClusterNodeEntryTest.java
@@ -1,5 +1,12 @@
 package org.cluster.membership.protocol;
 
+import java.util.TimeZone;
+
+import org.cluster.membership.common.model.Node;
+import org.cluster.membership.common.model.util.EnvUtils;
+import org.springframework.boot.ApplicationArguments;
+import org.springframework.boot.DefaultApplicationArguments;
+
 import junit.framework.Test;
 import junit.framework.TestCase;
 import junit.framework.TestSuite;
@@ -7,19 +14,25 @@ import junit.framework.TestSuite;
 /**
  * Unit test for simple App.
  */
-public class ClusterNodeEntryTest 
-    extends TestCase
-{
+public class ClusterNodeEntryTest extends TestCase {
     /**
      * Create the test case
      *
      * @param testName name of the test case
      */
-    public ClusterNodeEntryTest( String testName )
-    {
+    public ClusterNodeEntryTest( String testName ) throws Exception {
         super( testName );
+        
+        //String home = System.getProperty("user.dir");        
+        Node node = new Node("A", "localhost", 6001, 7001, TimeZone.getDefault());
+        
+        String argsS = EnvUtils.generateNodeCommandLineArguments(node, 0) + " --mode=DEBUG"; 
+        			
+        ApplicationArguments defaultAA = new DefaultApplicationArguments(argsS.split("\\s+")); 
+        Config.read(defaultAA);
+        
     }
-
+    
     /**
      * @return the suite of tests being tested
      */
@@ -27,12 +40,9 @@ public class ClusterNodeEntryTest
     {
         return new TestSuite( ClusterNodeEntryTest.class );
     }
-
-    /**
-     * Rigourous Test :-)
-     */
-    public void testApp()
-    {
-        assertTrue( true );
+    
+    public void testWork() {
+    	assert(true);
     }
+    
 }
