+++ protocol/src/test/java/org/cluster/membership/protocol/core/RandomServiceTest.java
@@ -3,20 +3,20 @@ package org.cluster.membership.protocol.core;
 import java.util.TimeZone;
 
 import org.cluster.membership.common.model.Node;
+import org.cluster.membership.protocol.ClusterNodeEntryTest;
 import org.cluster.membership.protocol.model.Message;
 import org.cluster.membership.protocol.structures.DList;
 import org.cluster.membership.protocol.structures.ValuePriorityEntry;
 import org.cluster.membership.protocol.structures.ValuePrioritySet;
 
 import junit.framework.Test;
-import junit.framework.TestCase;
 import junit.framework.TestSuite;
 
 /**
  * Unit test for simple App.
  */
 public class RandomServiceTest 
-    extends TestCase
+    extends ClusterNodeEntryTest
 {
 	
 	private DList nodes = new DList();
@@ -44,7 +44,7 @@ public class RandomServiceTest
      *
      * @param testName name of the test case
      */
-    public RandomServiceTest( String testName )
+    public RandomServiceTest( String testName ) throws Exception
     {
         super( testName );
