+++ protocol/src/test/java/org/cluster/membership/protocol/structures/ValuePrioritySetInsertMessageTest.java
@@ -4,19 +4,17 @@ import java.util.Iterator;
 import java.util.TimeZone;
 
 import org.cluster.membership.common.model.Node;
+import org.cluster.membership.protocol.ClusterNodeEntryTest;
 import org.cluster.membership.protocol.core.MessageType;
 import org.cluster.membership.protocol.model.Message;
 
 import junit.framework.Test;
-import junit.framework.TestCase;
 import junit.framework.TestSuite;
 
 /**
  * Unit test for simple App.
  */
-public class ValuePrioritySetInsertMessageTest 
-    extends TestCase
-{
+public class ValuePrioritySetInsertMessageTest extends ClusterNodeEntryTest {
 	
 	
 	private final Node a = new Node("A", "A 1", 7001, 6001, TimeZone.getDefault());
@@ -30,8 +28,7 @@ public class ValuePrioritySetInsertMessageTest
      *
      * @param testName name of the test case
      */
-    public ValuePrioritySetInsertMessageTest( String testName )
-    {
+    public ValuePrioritySetInsertMessageTest( String testName ) throws Exception {
         super( testName );
     }
