+++ tester/src/main/java/org/cluster/membership/tester/config/LocalEnvConfig.java
@@ -1,0 +1,9 @@
+package org.cluster.membership.tester.config;
+
+public class LocalEnvConfig extends AbstractEnvConfig {
+	
+	public LocalEnvConfig(String homePath, String programPath) throws Exception {
+		super(homePath, programPath);
+	}
+
+}
