+++ tester/src/main/java/org/cluster/membership/tester/config/MultipleVMEnvConfig.java
@@ -1,0 +1,10 @@
+package org.cluster.membership.tester.config;
+
+public class MultipleVMEnvConfig extends LocalEnvConfig {
+
+	
+	public MultipleVMEnvConfig(String homePath, String programPath) throws Exception {
+		super(homePath, programPath);
+	}
+
+}
