+++ tester/src/test/java/org/cluster/membership/tester/MultipleVMTesterEntryTest.java
@@ -1,4 +1,4 @@
-package org.cluster.membership.protocol.protocol;
+package org.cluster.membership.tester;
 
 import junit.framework.Test;
 import junit.framework.TestCase;
@@ -7,7 +7,7 @@ import junit.framework.TestSuite;
 /**
  * Unit test for simple App.
  */
-public class AppTest 
+public class MultipleVMTesterEntryTest 
     extends TestCase
 {
     /**
@@ -15,7 +15,7 @@ public class AppTest
      *
      * @param testName name of the test case
      */
-    public AppTest( String testName )
+    public MultipleVMTesterEntryTest( String testName )
     {
         super( testName );
     }
@@ -25,7 +25,7 @@ public class AppTest
      */
     public static Test suite()
     {
-        return new TestSuite( AppTest.class );
+        return new TestSuite( MultipleVMTesterEntryTest.class );
     }
 
     /**
@@ -33,6 +33,8 @@ public class AppTest
      */
     public void testApp()
     {
+    	String home = System.getProperty("user.dir");
+    	System.out.println(home);
         assertTrue( true );
     }
 }
